+++ comics/comics/whiteninja.py
@@ -1,4 +1,4 @@
-from comics.aggregator.crawler import CrawlerBase, CrawlerImage
+from comics.aggregator.crawler import CrawlerBase
 from comics.core.comic_data import ComicDataBase
 
 
@@ -7,19 +7,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.whiteninjacomics.com/'
     start_date = '2002-01-01'
+    end_date = '2012-08-04'
+    active = False
     rights = 'Scott Bevan & Kent Earle'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 60
-    time_zone = 'America/Regina'
-
     def crawl(self, pub_date):
-        feed = self.parse_feed(
-            'http://www.whiteninjacomics.com/rss/z-latest.xml')
-        for entry in feed.for_date(pub_date):
-            page = self.parse_page(entry.link)
-            page.remove('img[src*="/images/comics/t-"]')
-            url = page.src('img[src*="/images/comics/"]')
-            title = entry.title.split(' - ')[0]
-            return CrawlerImage(url, title)
+        pass  # Comic no longer published